abc2/abc/requirements.txt
flask==2.3.3
flask-cors==4.0.0
google-generativeai==0.3.2
python-dotenv==1.0.0
opencv-python==4.8.1.78
speechrecognition==3.10.0
pygame==2.5.2
boto3==1.29.7
pillow==10.0.1
pygetwindow==0.0.9
scipy==1.11.4
requests==2.31.0
